관리자 추가하여 관리자만 회원 목록 조회 가능

+++ src/main/java/com/example/pjjsp241022/controller/MemberController.java
@@ -33,10 +33,11 @@ public class MemberController {
 
         service.addMember(member);
 
-        rttr.addFlashAttribute("message", Map.of("type", "success",
+        rttr.addFlashAttribute("message", Map.of(
+                "type", "success",
                 "text", "회원가입 되었습니다."));
 
-        return "redirect:/member/list";
+        return "redirect:/board/list";
     }
 
     @GetMapping("list")
@@ -63,46 +64,80 @@ public class MemberController {
     }
 
     @PostMapping("delete")
-    public String deleteMember(String id, String password, RedirectAttributes rttr) {
+    public String deleteMember(String id,
+                               String password,
+                               HttpSession session,
+                               RedirectAttributes rttr,
+                               @SessionAttribute("loggedInMember") Member member) {
 
-        if (service.remove(id, password)) {
-            // 탈퇴 성공
-            rttr.addFlashAttribute("message", Map.of(
-                    "type", "warning",
-                    "text", "탈퇴가 완료되었습니다."));
-            return "redirect:/member/signup";
+        if (service.hashAccess(id, member)) {
+            if (service.remove(id, password)) {
+                // 탈퇴 성공
+                rttr.addFlashAttribute("message", Map.of(
+                        "type", "success",
+                        "text", "탈퇴가 완료되었습니다."));
+                session.invalidate();
+                return "redirect:/member/signup";
+            } else {
+                // 탈퇴 실패
+                rttr.addFlashAttribute("message", Map.of(
+                        "type", "danger",
+                        "text", "비밀번호가 일치하지 않습니다."));
+                rttr.addAttribute("id", id);
+                return "redirect:/member/view";
+            }
         } else {
-            // 탈퇴 실패
             rttr.addFlashAttribute("message", Map.of(
                     "type", "danger",
-                    "text", "비밀번호가 일치하지 않습니다."));
-            rttr.addAttribute("id", id);
-            return "redirect:/member/view";
+                    "text", "권한이 없습니다."));
+
+            session.invalidate();
+            return "redirect:/member/login";
         }
+
+
     }
 
     @GetMapping("edit")
-    public void edit(String id, Model model) {
-        model.addAttribute("member", service.info(id));
+    public String edit(String id, Model model,
+                       RedirectAttributes rttr,
+                       @SessionAttribute("loggedInMember") Member member) {
+
+        if (service.hasAccess(id, member)) {
+            model.addAttribute("member", service.info(id));
+            return null;
+        } else {
+            rttr.addFlashAttribute("message", Map.of("type", "danger",
+                    "text", "권한이 없습니다."));
+            return "redirect:/member/login";
+        }
     }
 
     @PostMapping("edit")
-    public String editProcess(Member member, RedirectAttributes rttr) {
-        try {
-            service.update(member);
-            rttr.addFlashAttribute("message", Map.of("type", "success",
-                    "text", "회원정보가 수정되었습니다."));
+    public String editProcess(Member member, RedirectAttributes rttr,
+                              @SessionAttribute("loggedInMember") Member loggedInMember) {
+        if (service.hasAccess(member.getId(), loggedInMember)) {
+            try {
+                service.update(member);
+                rttr.addFlashAttribute("message", Map.of("type", "success",
+                        "text", "회원정보가 수정되었습니다."));
 
-        } catch (DuplicateKeyException e) {
-            rttr.addFlashAttribute("message", Map.of("type", "danger",
-                    "text", STR."\{member.getNickName()}은 이미 사용중인 별명입니다."));
+            } catch (DuplicateKeyException e) {
+                rttr.addFlashAttribute("message", Map.of("type", "danger",
+                        "text", STR."\{member.getNickName()}은 이미 사용중인 별명입니다."));
+
+                rttr.addAttribute("id", member.getId());
+                return "redirect:/member/edit";
+            }
 
             rttr.addAttribute("id", member.getId());
-            return "redirect:/member/edit";
-        }
+            return "redirect:/member/view";
+        } else {
 
-        rttr.addAttribute("id", member.getId());
-        return "redirect:/member/view";
+            rttr.addFlashAttribute("message", Map.of("type", "danger",
+                    "text", "권한이 없습니다."));
+            return "redirect:/member/login";
+        }
     }
 
     @GetMapping("edit-password")
